[STATIC] sofa scraper
Made the static version of the sofa scraper, and also the file learn_class.py. This is me trying to make classes for hvitevarer and sofa making the code more dynamic, it needs alot of work to be done for.

Static version is almost ready, one flaw is that models/brands which has two words, fexp "Le Grande" cannot be scraped as of now, because we split words one by one. This needs to be looked at.

Furthermore, the live version of the scraper of sofa is not ready.

We alos need to update and go throug static V8 to fix the exeptions